LLama.Examples: RepoUtils.cs → ConsoleLogger.cs
